event-driven configuration support (#30)
Basic idea: make configuration across all nodes event-driven. We use a distributed KV store (can be started in-process for testing with --embed-master) which different components may register interest in. As changes are detected in their directories of interest, they are notified of changes by calling their Created/Modified/Deleted callbacks.

We also maintain a configuration cache of the interesting state, to reduce complexity of individual components. CachedGet may be used to access this cache. Non-cached gets may be issued with a normal Get command.