switch to libkv fork with etcd v3 support. It fixes race condition in test and improve general stability. Closes #222 (#317)
This commit switch to libkv for (serverless/libkv) which is based on abronan/libkv fork. It supports etcd v3. The watcher part was rewritten and simplified (no need to juggle with cache). 

This commit also removes reconciliation mechanism completely. libkv.WatchTree on etcd v3 is reliable enough.

Retry and backoff mechanism is handled by the etcd client internally.